fix: Auth.js type issues
